Add files via upload
This is the original data that was scraped off of Zappos. There are 2000 images of shoes and they all follow the same general format (one shoe behind one shoe in front). This data is used in the first version of the model and is a bit harder to work with because the model has to learn the second shoe as well. The second version of the model will be trained on shoes where the image is just one shoe.